refactor(bundler): move sign out of wix.rs
