+++ tooling/bundler/src/bundle/windows/mod.rs
@@ -1,1 +1,2 @@
 pub mod msi;
+pub mod sign;
